moved tests to sub-directories
added conv2d tests
